+++ packages/wrapper/web/src/localeStrategy/localeStrategyProvider.js
@@ -1,13 +1,13 @@
 export default ({ locale = 'en' } = { locale: 'en' }) => async (innerProps = {}) => {
   const mergedLocale = innerProps.locale || locale;
   if (mergedLocale === 'en') {
-    return innerProps;
+    return {};
   }
   try {
     const localeResource = await import(
       `wix-rich-content-common/dist/statics/locale/messages_${mergedLocale}.json`
     ).then(res => res.default);
-    return { locale: mergedLocale, localeResource, ...innerProps };
+    return { locale: mergedLocale, localeResource };
   } catch (err) {
     throw new Error(`error while loading locale ${locale}`, err);
   }