+++ packages/editor/web/src/lib/editorStateConversion.js
@@ -6,7 +6,21 @@ const addVersion = (obj, version) => {
   return obj;
 };
 
-const convertToRaw = ContentState => addVersion(toRaw(ContentState), version);
+const fixBlockDataImmutableJS = contentState => {
+  contentState.blocks.forEach(block =>
+    Object.keys(block.data).forEach(key => {
+      const value = block.data[key];
+      if (value.toObject) {
+        block.data[key] = value.toObject();
+      }
+    })
+  );
+  return contentState;
+};
+
+const convertToRaw = ContentState =>
+  addVersion(fixBlockDataImmutableJS(toRaw(ContentState)), version);
+
 const convertFromRaw = rawState => addVersion(fromRaw(rawState), rawState.VERSION);
 
 const createEmpty = () => addVersion(EditorState.createEmpty(), version);