+++ packages/wrapper/web/src/EngineWrapper.js
@@ -1,147 +1,63 @@
-import React, { Children, Suspense } from 'react';
+import React, { Children, Fragment } from 'react';
 import PropTypes from 'prop-types';
-import { modalStyles } from './themeStrategy/defaults';
+import FullscreenRenderer from './FullscreenRenderer';
+import ModalRenderer from './ModalRenderer';
+import { assert } from './utils';
 
 class EngineWrapper extends React.Component {
   constructor(props) {
     super(props);
-    this.state = {
-      disabled: false,
-      showModal: false,
-      EditorModal: undefined,
-      Fullscreen: undefined,
-    };
+    this.state = this.stateFromProps(props);
+    // BARAK: why dynamic?
     if (props.editor) {
       import(
         // eslint-disable-next-line max-len
         /* webpackChunkName: "rce-editorStateConversion"  */ `wix-rich-content-editor/dist/lib/editorStateConversion.js`
       ).then(module => this.setState({ editorState: module.createEmpty() }));
     }
-    this.emptyInitialState = { entityMap: {} };
   }
 
-  onModalOpen = data => {
-    const { modalStyles, ...modalProps } = data;
-    this.setState({
-      showModal: true,
-      modalProps,
-      modalStyles,
-    });
-    this.props.onModalOpen?.(data);
-  };
-
-  onModalClose = () => {
-    this.setState({
-      showModal: false,
-      modalProps: null,
-      modalStyles: null,
-      modalContent: null,
-    });
-    this.props.onModalClose?.();
-  };
+  stateFromProps(props) {
+    const { editor, children } = props;
+    const { closeModal, openModal, onExpand } = children.props?.helpers || {};
+    let ModalityProvider = Fragment;
+    if (editor && !closeModal && !openModal) {
+      ModalityProvider = ModalRenderer;
+    } else if (!editor && !onExpand) {
+      ModalityProvider = FullscreenRenderer;
+    }
+    return { ModalityProvider, disabled: false }; // BARAK: what this "disabled" state for?
+  }
 
   handleChange = editorState => {
     this.setState({ editorState });
   };
 
-  onExpand = (entityIndex, innerIndex = 0) => {
-    //galleries have an innerIndex (i.e. second image will have innerIndex=1)
-    this.setState({
-      //Viewer state
-      expandModeIsOpen: true,
-      expandModeIndex: this.state.expandModeData?.imageMap[entityIndex] + innerIndex,
-    });
-  };
-
-  setExpandModeData = expandModeData => {
-    this.setState({ expandModeData });
-  };
-
-  componentDidMount() {
-    const { editor, withModal = true } = this.props;
-    const shouldRenderEditorModal = editor && withModal;
-    const shouldRenderFullscreen = !editor && withModal;
-    let EditorModal, Fullscreen;
-    if (shouldRenderEditorModal)
-      EditorModal =
-        withModal === true
-          ? React.lazy(() => import(/* webpackChunkName: "rce-EditorModal"  */ `./EditorModal.js`))
-          : withModal;
-    if (shouldRenderFullscreen) {
-      Fullscreen =
-        withModal === true
-          ? React.lazy(() => import(/* webpackChunkName: "rce-ViewerModal"  */ './ViewerModal'))
-          : withModal;
-    }
-    this.setState({ EditorModal, Fullscreen });
-  }
-
   render() {
-    const { strategies = [], children = {}, withModal = true, editor } = this.props;
-    const modifiedProps = strategies.reduce(
-      (props, strategyFunction) => {
-        const result = strategyFunction(props);
-        return { ...props, ...result };
-      },
-      { ...children.props }
+    const { strategies = [], children, editor } = this.props;
+
+    const childProps = strategies.reduce(
+      (props, strategy) => ({ ...props, ...strategy(props) }),
+      children.props
     );
-    const { helpers = {}, theme, locale = 'en', ModalsMap, onChange } = modifiedProps;
-    const { onRequestClose } = this.state.modalProps || {};
 
-    //viewer needs onExpand helper + Fullscreen
-    helpers.onExpand = this.onExpand;
+    assert(childProps.theme && childProps.config, '[EngineWrapper] invalid strategy reduce');
 
-    //Editor Modal
-    if (editor && withModal) {
-      helpers.openModal = this.onModalOpen;
-      helpers.closeModal = this.onModalClose;
+    // BARAK: why do we need this?
+    if (editor) {
+      const { onChange } = childProps;
+      childProps.onChange = editorState => {
+        onChange?.(editorState);
+        this.handleChange(editorState);
+      };
     }
-    modifiedProps.onChange = editorState => {
-      onChange?.(editorState);
-      this.handleChange(editorState);
-    };
 
-    modifiedProps.helpers = helpers;
+    const { disabled, ModalityProvider } = this.state;
 
-    const {
-      expandModeData,
-      expandModeIndex,
-      expandModeIsOpen,
-      disabled,
-      EditorModal,
-      Fullscreen,
-    } = this.state;
     return (
-      <React.Fragment>
-        <div id="#wrapper_viewer_modal" />
-        {Children.only(React.cloneElement(children, { ...modifiedProps, disabled }))}
-        {EditorModal && (
-          <Suspense fallback={<div />}>
-            <EditorModal
-              isOpen={this.state.showModal}
-              contentLabel="External Modal Example"
-              style={modalStyles(this.state, theme)}
-              role="dialog"
-              onRequestClose={onRequestClose || helpers.closeModal}
-              modalsMap={ModalsMap}
-              locale={locale}
-              {...this.state.modalProps}
-            />
-          </Suspense>
-        )}
-        {Fullscreen && (
-          <Suspense fallback={<div />}>
-            <Fullscreen
-              initialState={children.props.initialState || this.emptyInitialState}
-              isOpen={expandModeIsOpen}
-              images={expandModeData?.images || []}
-              onClose={() => this.setState({ expandModeIsOpen: false })}
-              index={expandModeIndex}
-              setExpandModeData={this.setExpandModeData}
-            />
-          </Suspense>
-        )}
-      </React.Fragment>
+      <ModalityProvider {...childProps}>
+        {Children.only(React.cloneElement(children, { ...childProps, disabled }))}
+      </ModalityProvider>
     );
   }
 }
@@ -149,11 +65,7 @@ EngineWrapper.propTypes = {
   strategies: PropTypes.array,
   plugins: PropTypes.arrayOf(PropTypes.object),
   theme: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
-  onRequestModalClose: PropTypes.func,
-  onModalOpen: PropTypes.func,
-  onModalClose: PropTypes.func,
   children: PropTypes.object,
   editor: PropTypes.bool,
-  withModal: PropTypes.oneOf([PropTypes.bool, PropTypes.object]),
 };
 export default EngineWrapper;