+++ packages/wrapper/web/src/themeStrategy/ThemeGenerator.js
@@ -86,6 +86,15 @@ export default class ThemeGenerator {
             fill: actionColor,
           },
         },
+        footerToolbar: {
+          background: bgColor,
+        },
+        footerToolbarButton_icon: {
+          color: textColor,
+          ':hover': {
+            color: actionColor,
+          },
+        },
         //block focus
         hasFocus: blockActionColorSettings,
         pluginContainer: { ':hover': blockActionColorSettings },