+++ plugins-bundle-analyzer/src/editorStateConversion.js
@@ -1,0 +1,4 @@
+import editorStateConversion from 'wix-rich-content-editor/dist/lib/editorStateConversion.cjs';
+
+console.log({ editorStateConversion }); //eslint-disable-line
+export default () => {};